requirements.txt
langgraph>=0.2.39
langchain-core>=0.3.19
langchain>=0.3.7
langchain-community>=0.3.7
langchain-openai>=0.2.9
python-dotenv>=1.0.1
gradio>=4.44.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.8.0

